made a util class for mathematical functions
